geometry refactor in future
