+++ DeepINN/geometry/__init__.py
@@ -1,1 +1,5 @@
-pass
+# Make the subdirectory callable
+from . import domains
+#from . import conditions
+from . import samplers
+from . import spaces
