+++ DeepINN/geometry/domains/domain2D/shapely_polygon.py
