+++ DeepINN/geometry/domains/domain2D/triangle.py
