+++ DeepINN/geometry_refactor/spaces/__init__.py
@@ -1,0 +1,1 @@
+from .space import Space, R1, R2, R3
