+++ DeepINN/model.py
@@ -26,7 +26,7 @@ class Model():
         self.compile_network()
 
     def compile_domain(self):
-        # sample collcation points
+        # sample collocation points
         self.collocation_point_sample, self.collocation_point_labels = self.domain.sample_collocation_labels()
 
         # sample boundary points
@@ -54,7 +54,7 @@ class Model():
             self.config.apply_float_type()
             self.config.default_device()
 
-            # In 1D problem we need to combine the BCs as there is only one point for each BC, which returns an underfined feature scaling because the ub and lb are same in the denominator
+            # In 1D problem we need to combine the BCs as there is only one point for each BC, which returns an undefined feature scaling because the ub and lb are same in the denominator, so we get infinity
             # For problem with multiple points on each boundary, we don't need to combine them.
             if self.boundary_point_sample[0].size()[0] == 1: # if row is 1 in the particular boundary tensor
                 self.boundary_point_sample = torch.cat(self.boundary_point_sample, dim=0)
@@ -80,7 +80,7 @@ class Model():
             # backprop the total loss
             self.total_loss.backward() 
             
-            # Update model parameters based on the older values and the backproped gradient
+            # Update model parameters based on the older values and the backprop gradient
             self.optimiser.step()
             if self.iter % (self.iterations/10) == 0:
                 print(f"Iteration: {self.iter+1} \t BC Loss: {self.BC_loss:0.4f}\t PDE Loss: {self.PDE_loss:0.4f} \t Loss: {self.total_loss:0.4f}")
